# packages/ropa-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RoPA API",
    "version": "0.1.0",
    "description": "Record of Processing Activities (GDPR Article 30).\n\nReads are public by default and an anonymous caller is treated as a viewer; every write needs a token. Mint one at `POST /v1/tokens`, then use **Authorize** above.\n\nDemo project. All data is synthetic; no real personal data is used."
  },
  "servers": [
    {
      "url": "/",
      "description": "This service"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A token from POST /v1/tokens. Paste the token itself, without \"Bearer\"."
      }
    },
    "schemas": {
...
      "DataCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "A UUID. Records are keyed by UUIDv7."
          },
          "version": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "An RFC 3339 timestamp."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "An RFC 3339 timestamp."
          },
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "description": "A human-readable identifier, unique within its table: \"mailcrest\", \"standard-dpa-v3\"."
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ]
          },
          "special": {
            "type": "string",
            "enum": [
              "none",
              "art9",
              "art10"
            ],
            "description": "art9: a special category. art10: criminal convictions."
          }
        },
        "required": [
          "id",
          "version",
          "createdAt",
          "updatedAt",
          "slug",
          "name",
          "description",
          "special"
        ],
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "code": {
                  "type": "string",
                  "minLength": 1
                },
                "message": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "required": [
                "path",
                "code",
                "message"
              ],
              "additionalProperties": false,
              "description": "One field-level validation failure."
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": {},
        "description": "An RFC 9457 problem document (application/problem+json)."
      },
...
      "DataCategoryInput": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "description": "A human-readable identifier, unique within its table: \"mailcrest\", \"standard-dpa-v3\"."
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "changeNote": {
            "description": "Why this change was made. Stored on the revision, never returned.",
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "special": {
            "default": "none",
            "type": "string",
            "enum": [
              "none",
              "art9",
              "art10"
            ],
            "description": "art9: a special category. art10: criminal convictions."
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  },
  "paths": {
...
    "/v1/taxonomy/data-categories": {
      "get": {
        "tags": [
          "data-categories"
        ],
        "summary": "List data category records",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Page size (default 50, max 200).",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "From a previous page.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "nextCursor"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DataCategory"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "description": "Pass as ?cursor= for the next page. Null on the last page."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The limit or cursor is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Valid token without the required permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "description": "Requires the `record:read` permission.",
        "x-required-permission": "record:read"
      },
      "post": {
        "tags": [
          "data-categories"
        ],
        "summary": "Create a data category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataCategoryInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataCategory"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "The record version. Send it back as If-Match on a write (§1.8).",
                "schema": {
                  "type": "string",
                  "example": "\"3\""
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Valid token without the required permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "An identifier is already taken",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "description": "Requires the `taxonomy:write` permission.",
        "x-required-permission": "taxonomy:write"
      }
    },
    "/v1/taxonomy/data-categories/{ref}": {
      "get": {
        "tags": [
          "data-categories"
        ],
        "summary": "Read a data category",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "The data category: an id, code or slug (§1.2).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataCategory"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "The record version. Send it back as If-Match on a write (§1.8).",
                "schema": {
                  "type": "string",
                  "example": "\"3\""
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Valid token without the required permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No record matching that identifier",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "description": "Requires the `record:read` permission.",
        "x-required-permission": "record:read"
      },
      "put": {
        "tags": [
          "data-categories"
        ],
        "summary": "Replace a data category",
        "description": "A full replacement, not a patch: each revision stays a clean snapshot (§1.4).\n\nRequires the `taxonomy:write` permission.",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "The data category: an id, code or slug (§1.2).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The version this write is based on, from the record’s ETag (§1.8).",
            "schema": {
              "type": "string",
              "example": "\"3\""
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataCategoryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The saved record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataCategory"
                }
              }
            },
            "headers": {
              "ETag": {
                "description": "The record version. Send it back as If-Match on a write (§1.8).",
                "schema": {
                  "type": "string",
                  "example": "\"3\""
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Valid token without the required permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No record matching that identifier",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Someone else has saved since you read this record",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "If-Match is required on a write",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "x-required-permission": "taxonomy:write"
      },
      "delete": {
        "tags": [
          "data-categories"
        ],
        "summary": "Delete a data category",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "The data category: an id, code or slug (§1.2).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "description": "The version this write is based on, from the record’s ETag (§1.8).",
            "schema": {
              "type": "string",
              "example": "\"3\""
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted. Its history remains."
          },
          "401": {
            "description": "No token, or the token is not valid",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Valid token without the required permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No record matching that identifier",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Something still references this record",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "412": {
            "description": "Someone else has saved since you read this record",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "The request failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "428": {
            "description": "If-Match is required on a write",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "description": "Requires the `record:delete` permission.",
        "x-required-permission": "record:delete"
      }
    },
...
  }
}